[REFACTOR] 유저 모듈 구조 변경
