+++ src/user/entity/user-account.entity.ts
@@ -7,8 +7,8 @@ import {
   OneToOne,
   JoinColumn,
 } from 'typeorm';
-import { UserProfileEntity } from '@src/user/common/entity/user-profile.entity';
-import { AccountStatus } from '@src/user/kakao/types/account-status';
+import { UserProfileEntity } from '@src/user/entity/user-profile.entity';
+import { AccountStatus } from '@src/user/types/account-status';
 
 @Entity('account')
 export class UserAccountEntity {
