--- src/user/google/test/google.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { GoogleService } from '../google-user.service';
-
-describe('GoogleService', () => {
-  let service: GoogleService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [GoogleService],
-    }).compile();
-
-    service = module.get<GoogleService>(GoogleService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
