+++ src/user/user.repository.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { DataSource, Repository } from 'typeorm';
-import { UserProfileEntity } from '@src/user/common/entity/user-profile.entity';
-import { UserAccountEntity } from '@src/user/common/entity/user-account.entity';
-import { AccountStatus } from '@src/user/kakao/types/account-status';
+import { UserProfileEntity } from '@src/user/entity/user-profile.entity';
+import { UserAccountEntity } from '@src/user/entity/user-account.entity';
+import { AccountStatus } from '@src/user/types/account-status';
 
 @Injectable()
 export class UserRepository {
